Converted NeuronGPUTime to double, NeuronGPUTimeIdx, LastSpikeTimeIdx and LastRevSpikeTimeIdx to long long

+++ src/aeif_cond_beta_kernel.h
@@ -137,7 +137,7 @@ const std::string aeif_cond_beta_port_param_name[N_PORT_PARAM] = {
 
  template<int NVAR, int NPARAM> //, class DataStruct>
 __device__
-    void Derivatives(float x, float *y, float *dydx, float *param,
+    void Derivatives(double x, float *y, float *dydx, float *param,
 		     aeif_cond_beta_rk5 data_struct)
 {
   enum { n_port = (NVAR-N_SCAL_VAR)/N_PORT_VAR };
@@ -163,7 +163,7 @@ __device__
  template<int NVAR, int NPARAM> //, class DataStruct>
 __device__
     void ExternalUpdate
-    (float x, float *y, float *param, bool end_time_step,
+    (double x, float *y, float *param, bool end_time_step,
 			aeif_cond_beta_rk5 data_struct)
 {
   if ( V_m < -1.0e3) { // numerical instability
@@ -202,10 +202,10 @@ __device__
 };
 
 template <>
-int aeif_cond_beta::UpdateNR<0>(int it, float t1);
+int aeif_cond_beta::UpdateNR<0>(long long it, double t1);
 
 template<int N_PORT>
-int aeif_cond_beta::UpdateNR(int it, float t1)
+int aeif_cond_beta::UpdateNR(long long it, double t1)
 {
   if (N_PORT == n_port_) {
     const int NVAR = aeif_cond_beta_ns::N_SCAL_VAR
@@ -224,7 +224,7 @@ int aeif_cond_beta::UpdateNR(int it, float t1)
 
 template<int NVAR, int NPARAM>
 __device__
-void Derivatives(float x, float *y, float *dydx, float *param,
+void Derivatives(double x, float *y, float *dydx, float *param,
 		 aeif_cond_beta_rk5 data_struct)
 {
     aeif_cond_beta_ns::Derivatives<NVAR, NPARAM>(x, y, dydx, param,
@@ -233,7 +233,7 @@ void Derivatives(float x, float *y, float *dydx, float *param,
 
 template<int NVAR, int NPARAM>
 __device__
-void ExternalUpdate(float x, float *y, float *param, bool end_time_step,
+void ExternalUpdate(double x, float *y, float *param, bool end_time_step,
 		    aeif_cond_beta_rk5 data_struct)
 {
     aeif_cond_beta_ns::ExternalUpdate<NVAR, NPARAM>(x, y, param,
