+++ src/ext_neuron.h
@@ -68,9 +68,9 @@ class ext_neuron : public BaseNeuron
   int Init(int i_node_0, int n_neuron, int n_port, int i_group,
 	   unsigned long long *seed);
 
-  //int Calibrate(float time_min, float time_resolution);
+  //int Calibrate(double time_min, float time_resolution);
 		
-  int Update(int it, float t1);
+  int Update(long long it, double t1);
 
   int Free();
