+++ src/iaf_psc_exp_hc.h
@@ -62,7 +62,7 @@ class iaf_psc_exp_hc : public BaseNeuron
   int Init(int i_node_0, int n_neuron, int n_port, int i_group,
 	   unsigned long long *seed);
 
-  int Update(int it, float t1);
+  int Update(long long it, double t1);
 
   int Free();
